+++ wrappers/golang/curves/bls12377/bls12_377.go
@@ -1,4 +1,4 @@
 package bls12377
 
-// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build -lingo_bls12_377 -lstdc++ -lm
+// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build/src/curves -L${SRCDIR}/../../../../icicle/build/src/fields -lingo_curve_bls12_377 -lingo_field_bls12_377 -lstdc++ -lm
 import "C"
