+++ wrappers/golang/curves/bls12381/bls12_381.go
@@ -1,4 +1,4 @@
 package bls12381
 
-// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build -lingo_bls12_381 -lstdc++ -lm
+// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build/src/curves -L${SRCDIR}/../../../../icicle/build/src/fields -lingo_curve_bls12_381 -lingo_field_bls12_381 -lstdc++ -lm
 import "C"
