+++ wrappers/golang/curves/bn254/bn254.go
@@ -1,4 +1,4 @@
 package bn254
 
-// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build -lingo_bn254 -lstdc++ -lm
+// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build/src/curves -L${SRCDIR}/../../../../icicle/build/src/fields -lingo_curve_bn254 -lingo_field_bn254 -lstdc++ -lm
 import "C"
