+++ wrappers/golang/curves/bn254/ntt.go
@@ -53,33 +53,6 @@ func Ntt[T any](scalars core.HostOrDeviceSlice, dir core.NTTDir, cfg *core.NTTCo
 	return core.FromCudaError(err)
 }
 
-func ECNtt[T any](points core.HostOrDeviceSlice, dir core.NTTDir, cfg *core.NTTConfig[T], results core.HostOrDeviceSlice) core.IcicleError {
-	core.NttCheck[T](points, cfg, results)
-
-	var pointsPointer unsafe.Pointer
-	if points.IsOnDevice() {
-		pointsPointer = points.(core.DeviceSlice).AsPointer()
-	} else {
-		pointsPointer = unsafe.Pointer(&points.(core.HostSlice[Projective])[0])
-	}
-	cPoints := (*C.projective_t)(pointsPointer)
-	cSize := (C.int)(points.Len() / int(cfg.BatchSize))
-	cDir := (C.int)(dir)
-	cCfg := (*C.NTTConfig)(unsafe.Pointer(cfg))
-
-	var resultsPointer unsafe.Pointer
-	if results.IsOnDevice() {
-		resultsPointer = results.(core.DeviceSlice).AsPointer()
-	} else {
-		resultsPointer = unsafe.Pointer(&results.(core.HostSlice[Projective])[0])
-	}
-	cResults := (*C.projective_t)(resultsPointer)
-
-	__ret := C.bn254ECNTTCuda(cPoints, cSize, cDir, cCfg, cResults)
-	err := (cr.CudaError)(__ret)
-	return core.FromCudaError(err)
-}
-
 func InitDomain(primitiveRoot ScalarField, ctx cr.DeviceContext, fastTwiddles bool) core.IcicleError {
 	cPrimitiveRoot := (*C.scalar_t)(unsafe.Pointer(primitiveRoot.AsPointer()))
 	cCtx := (*C.DeviceContext)(unsafe.Pointer(&ctx))
