+++ wrappers/golang/curves/bw6761/bw6_761.go
@@ -1,4 +1,4 @@
 package bw6761
 
-// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build -lingo_bw6_761 -lstdc++ -lm
+// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build/src/curves -L${SRCDIR}/../../../../icicle/build/src/fields -lingo_curve_bw6_761 -lingo_field_bw6_761 -lstdc++ -lm
 import "C"
