+++ wrappers/golang/internal/generator/templates/main.go.tmpl
@@ -1,4 +1,4 @@
 package {{.PackageName}}
 
-// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build -lingo_{{.Curve}} -lstdc++ -lm
+// #cgo LDFLAGS: -L${SRCDIR}/../../../../icicle/build/src/curves -L${SRCDIR}/../../../../icicle/build/src/fields -lingo_curve_{{.Curve}} -lingo_field_{{.Curve}} -lstdc++ -lm
 import "C"
