+++ wrappers/golang/internal/generator/templates/ntt_test.go.tmpl
@@ -106,25 +106,6 @@ func TestNtt(t *testing.T) {
 	}
 }
 
-func TestECNtt(t *testing.T) {
-	cfg := GetDefaultNttConfig()
-	points := GenerateProjectivePoints(1 << largestTestSize)
-
-	for _, size := range []int{4, 5, 6, 7, 8} {
-		for _, v := range [4]core.Ordering{core.KNN, core.KNR, core.KRN, core.KRR} {
-			testSize := 1 << size
-
-			pointsCopy := core.HostSliceFromElements[Projective](points[:testSize])
-			cfg.Ordering = v
-			cfg.NttAlgorithm = core.Radix2
-
-			output := make(core.HostSlice[Projective], testSize)
-			e := ECNtt(pointsCopy, core.KForward, &cfg, output)
-			assert.Equal(t, core.IcicleErrorCode(0), e.IcicleErrorCode, "ECNtt failed")
-		}
-	}
-}
-
 func TestNttDeviceAsync(t *testing.T) {
 	cfg := GetDefaultNttConfig()
 	scalars := GenerateScalars(1 << largestTestSize)
